pyproject.toml
[project]
name = "rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "qdrant-client>=1.16.1",
    "uvicorn>=0.38.0",
    "google-generativeai>=0.8.3",
]
